requirements.txt
django==5.1.1
python-telegram-bot==21.5
django-rest-framework==3.15.2 
gunicorn
whitenoise